test: base api test 파일 분리
